scripts/requirements.txt
azure-identity==1.25.1
azure-storage-blob==12.27.1
Faker==38.2.0
snowflake-connector-python==4.1.1
colorlog==6.10.1